Change smallint -> integer

+++ migrations/189_workspace_hosts__load_count__add.sql
@@ -1,1 +1,1 @@
-ALTER TABLE workspace_hosts ADD COLUMN IF NOT EXISTS load_count smallint default 0;
+ALTER TABLE workspace_hosts ADD COLUMN IF NOT EXISTS load_count integer default 0;
